requirements.txt
django
aiogram
djangorestframework
environs~=8.0.0
requests
